+++ azext_iot/sdk/iothub/service/operations/bulk_registry_operations.py
@@ -23,7 +23,7 @@ class BulkRegistryOperations(object):
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
     :param deserializer: An object model deserializer.
-    :ivar api_version: Version of the Api. Constant value: "2021-04-12".
+    :ivar api_version: Version of the Api. Constant value: "2024-03-31".
     """
 
     models = models
@@ -33,7 +33,7 @@ class BulkRegistryOperations(object):
         self._client = client
         self._serialize = serializer
         self._deserialize = deserializer
-        self.api_version = "2021-04-12"
+        self.api_version = "2024-03-31"
 
         self.config = config
 
@@ -45,7 +45,7 @@ class BulkRegistryOperations(object):
         different devices are allowed. A maximum of 100 devices can be
         specified per invocation. For large scale operations, use the import
         feature using blob storage
-        (https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-identity-registry#import-and-export-device-identities).
+        (https://learn.microsoft.com/azure/iot-hub/iot-hub-devguide-identity-registry#import-and-export-device-identities).
 
         :param devices: The registry operations to perform.
         :type devices: list[~service.models.ExportImportDevice]